% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{App}
\alias{App}
\title{App}
\description{
This class represents a spatial app, for example a mobile app that you
might have installed on your smartphone.
}
\section{Usage}{

\preformatted{x = read.spu("dummy_data.csv")
app <- App$new("My app", x)

print(app)
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{process} object.}
  \item{name}{Character scalar, the name of the app.}
  \item{usage_data}{An sf object, representing the usage data for the app.}
}
}

\section{Details}{

\code{$new()} sets up the class with two fields: name and usage_data

\code{$print()} prints out some basic identifying information about the app.
}

\examples{
x = read.spu("dummy_data.csv")
app <- App$new("My app", x)

}
