% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Get username and password}
\usage{
get_credentials()
}
\value{
The username and password entered
}
\description{
Defaults to get from environment variable, falling back to getting them
interactively if supported
}
\keyword{internal}
