% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_client.R
\name{parse_security_token_response}
\alias{parse_security_token_response}
\title{Parse response from security token request}
\usage{
parse_security_token_response(response)
}
\arguments{
\item{response}{httr response object}
}
\value{
The security token or NA if failed to retrieve
}
\description{
This takes the full response and pulls out the part of the xml containing
the security token.
}
\keyword{internal}
