% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spu.R
\name{plot_usage_actions}
\alias{plot_usage_actions}
\title{Draw map showing spatial distribution of usage actions using mapview}
\usage{
plot_usage_actions(data)
}
\arguments{
\item{data}{An spu oject containing the data to be displayed}
}
\description{
Draw map showing spatial distribution of usage actions using mapview
}
