% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spu.R
\name{plot_user_path}
\alias{plot_user_path}
\title{Draw map showing a path of a user over time}
\usage{
plot_user_path(data, u)
}
\arguments{
\item{data}{An spu object containing the data to be displayed}

\item{u}{The identifier of one user}
}
\description{
Draw map showing a path of a user over time
}
