% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointr.R
\name{pointr}
\alias{pointr}
\title{Create sharepoint connection for downloading data.}
\description{
Create sharepoint connection for downloading data.

Create sharepoint connection for downloading data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pointr$new()}}
\item \href{#method-download}{\code{pointr$download()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create pointr object for downloading data from sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pointr$new(sharepoint_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_url}}{Root URL of sharepoint site to download from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pointr} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\subsection{Method \code{download()}}{
Download data from sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pointr$download(sharepoint_path, save_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_path}}{Path to the resource within sharepoint}

\item{\code{save_path}}{Path to save downloaded data to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Path to saved data
}
}
}
