% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_client.R
\name{prepare_security_token_payload}
\alias{prepare_security_token_payload}
\title{Prepare payload for retrieving security token}
\usage{
prepare_security_token_payload(url, credentials)
}
\arguments{
\item{url}{URL for site you are requesting a token for}

\item{credentials}{Username and password for site request token}
}
\value{
Formatted xml body for security token request
}
\description{
Prepare payload for retrieving security token
}
\keyword{internal}
