% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{read.spud_app}
\alias{read.spud_app}
\title{Read data from csv files directly into an App object.}
\usage{
read.spud_app(file, crs = 4326, name)
}
\arguments{
\item{file}{The name of the file you want to read in.}

\item{crs}{The coordinate reference system of the data.}

\item{name}{The name of the app.}
}
\value{
The resulting App object.
}
\description{
Read data from csv files directly into an App object.
}
