% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{read.spud_user}
\alias{read.spud_user}
\title{Read data from csv files directly into a User object.}
\usage{
read.spud_user(file, crs = 4326, id)
}
\arguments{
\item{file}{The name of the file you want to read in.}

\item{crs}{The coordinate reference system of the data.}

\item{id}{The id of the user.}
}
\value{
The resulting User object.
}
\description{
Read data from csv files directly into a User object.
}
