% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{sharepoint}
\alias{sharepoint}
\title{Create sharepoint connection for downloading data.}
\description{
Create sharepoint connection for downloading data.

Create sharepoint connection for downloading data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{A sharepoint_folder object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint$new()}}
\item \href{#method-download}{\code{sharepoint$download()}}
\item \href{#method-folder}{\code{sharepoint$folder()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create sharepoint object for downloading data from sharepoint. If
\code{auth} object provided any individual auth args will be used to
override those on the \code{auth} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$new(
  site_name = NULL,
  site_url = NULL,
  site_id = NULL,
  tenant = NULL,
  app = NULL,
  scopes = NULL,
  auth = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_name}}{The name of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}. If \code{NULL} will use env var
\code{SHAREPOINT_SITE_NAME}.}

\item{\code{site_url}}{The url of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}. If \code{NULL} will use env var
\code{SHAREPOINT_SITE_URL}.}

\item{\code{site_id}}{The ID of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}. If \code{NULL} will use env var
\code{SHAREPOINT_SITE_ID}.}

\item{\code{tenant}}{The name of the Azure Active Directory tenant
passed to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}. If \code{NULL} will use
env var \code{SHAREPOINT_SITE_TENANT}.}

\item{\code{app}}{A custom app registration ID to use for authentication
passed to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}. If \code{NULL} will use
env var \code{SHAREPOINT_SITE_APP_ID}.}

\item{\code{scopes}}{Microsoft graph scopes to obtain passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}.}

\item{\code{auth}}{\code{sharepoint_auth} object containing authentication data
passed to the client.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{sharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download data from sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$download(sharepoint_path, dest = NULL, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_path}}{Path to the resource within sharepoint}

\item{\code{dest}}{Path to save downloaded data to}

\item{\code{overwrite}}{Overwrite existing files?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Path to saved data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\if{latex}{\out{\hypertarget{method-folder}{}}}
\subsection{Method \code{folder()}}{
Create a \code{sharepoint_folder} object representing a sharepoint
folder, with which one can list, download and upload files.  See
\code{\link{sharepoint_folder}} for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$folder(path = "/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to folder from root of sharepoint site. Defaults to
"/" for the root.}
}
\if{html}{\out{</div>}}
}
}
}
