% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_client.R
\name{sharepoint_client}
\alias{sharepoint_client}
\title{Create sharepoint client, to manage HTTP requests to sharepoint.}
\description{
Create sharepoint client, to manage HTTP requests to sharepoint.

Create sharepoint client, to manage HTTP requests to sharepoint.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint_client$new()}}
\item \href{#method-GET}{\code{sharepoint_client$GET()}}
\item \href{#method-POST}{\code{sharepoint_client$POST()}}
\item \href{#method-request}{\code{sharepoint_client$request()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create client object for sending http requests to sharepoint.

This manages authenticating with sharepoint via sending credentails to
microsoft to retrieve an access token which it then sends to sharepoint
to retrieve cookies used for subsequent authentication.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_client$new(sharepoint_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_url}}{Root URL of sharepoint site to download from}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{sharepoint_client} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET"></a>}}
\subsection{Method \code{GET()}}{
Send GET request to sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_client$GET(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HTTP response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-POST"></a>}}
\subsection{Method \code{POST()}}{
Send POST request to sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_client$POST(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
HTTP response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\subsection{Method \code{request()}}{
Send POST request to sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_client$request(verb, path, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verb}}{A httr function for type of request to send e.g. httr::GET}

\item{\code{path}}{Request path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
HTTP response
}
}
}
