% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{sharepoint_download}
\alias{sharepoint_download}
\title{Download a dataset from sharepoint}
\usage{
sharepoint_download(
  sharepoint_path,
  dest = NULL,
  overwrite = FALSE,
  site_name = NULL,
  site_url = NULL,
  site_id = NULL,
  tenant = NULL,
  app = NULL,
  scopes = NULL
)
}
\arguments{
\item{sharepoint_path}{The path to the dataset you want to download - this
should include any subsites in the url and should be of the form
sites/nested/subsites/docs/path/to/document
e.g. if you want to get the file at Data/shape files/example.geojson from
the site groupA which is in site facultyA the full path would be
sites/facultyA/groupA/docs/Data/shape files/example.geojson
You should be able to get this if you locate the data you want to download
in a browser and click menu on the RHS of the file name which appears on
hover -> Copy link and manually edit to get the file path. See vignette for
more details.}

\item{dest}{Path to location you want to save the data. The default
save location is a tempfile with the same file extension as the downloaded
file.}

\item{overwrite}{if \code{TRUE} then the \code{dest} will be
ovewritten if it exists (otherwise it an error will be thrown)}

\item{site_name}{The name of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{site_url}{The url of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{site_id}{The ID of the sharepoint site passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{tenant}{The name of the Azure Active Directory tenant
passed to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{app}{A custom app registration ID to use for authentication
passed to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}

\item{scopes}{Microsoft graph scopes to obtain passed to
\code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}}
}
\value{
Path to downloaded data
}
\description{
Download a dataset from sharepoint
}
