% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_folder.R
\name{sharepoint_folder}
\alias{sharepoint_folder}
\title{Interact with sharepoint folders and their files.}
\description{
Interact with sharepoint folders and their files.

Interact with sharepoint folders and their files.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{site}}{Name of the sharepoint site (readonly)}

\item{\code{path}}{Path of the folder (readonly)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint_folder$new()}}
\item \href{#method-files}{\code{sharepoint_folder$files()}}
\item \href{#method-folders}{\code{sharepoint_folder$folders()}}
\item \href{#method-list}{\code{sharepoint_folder$list()}}
\item \href{#method-delete}{\code{sharepoint_folder$delete()}}
\item \href{#method-parent}{\code{sharepoint_folder$parent()}}
\item \href{#method-folder}{\code{sharepoint_folder$folder()}}
\item \href{#method-create}{\code{sharepoint_folder$create()}}
\item \href{#method-download}{\code{sharepoint_folder$download()}}
\item \href{#method-upload}{\code{sharepoint_folder$upload()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create sharepoint_folder object to enable listing, creating
downloading and uploading files & folders
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$new(client, site, path, verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{A low-level sharepoint client object, which can be used to
interact directly with the sharepoint API.}

\item{\code{site}}{The name of the sharepoint site (most likely a short string)}

\item{\code{path}}{Relative path within that shared site.  It seems
that "Shared Documents" is a common path that most likely
represents a "Documents" collection when viewed in the
sharepoint web interface.}

\item{\code{verify}}{Logical, indicating if the site/path combination is
valid (slower but safer).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-files"></a>}}
\if{latex}{\out{\hypertarget{method-files}{}}}
\subsection{Method \code{files()}}{
List all files within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$files(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folders"></a>}}
\if{latex}{\out{\hypertarget{method-folders}{}}}
\subsection{Method \code{folders()}}{
List all folders within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folders(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
List all folders and files within the folder; this is a
convenience wrapper around the \code{files} and \code{folders} methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$list(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a folder. Be extremely careful as you could
use this to delete an entire sharepoint. Deleted files are sent
to the recycle bin, so can be restored with relative ease, but
it will still be alarming. There is a mechanism to prevent
accidental deletion by declaring a file that exists within the
folder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$delete(path, check)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to delete. Use \code{NULL} to delete the current
folder.}

\item{\code{check}}{A file (not folder) that exists directly within
\code{path}, used as a method to verify that you really do want
to delete this folder (to prevent things like accidental deletion
of the entire sharepoint, for example).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parent"></a>}}
\if{latex}{\out{\hypertarget{method-parent}{}}}
\subsection{Method \code{parent()}}{
Create an object referring to the parent folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$parent(verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verify}}{Verify that the folder exists (which it must really here)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\if{latex}{\out{\hypertarget{method-folder}{}}}
\subsection{Method \code{folder()}}{
Create an object referring to a child folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folder(path, verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the folder, relative to this folder}

\item{\code{verify}}{Verify that the folder exists (which it must really here)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
Create a folder on sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$create(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download a file from a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$download(
  path,
  dest = NULL,
  progress = FALSE,
  overwrite = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the path to download, relative to this folder}

\item{\code{dest}}{Path to save downloaded data to. If \code{NULL} then a
temporary file with the same file extension as \code{path} is used.
If code{raw()} (or any other raw value) then the raw bytes will be
returned.}

\item{\code{progress}}{Display httr's progress bar?}

\item{\code{overwrite}}{Overwrite the file if it exists?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload a file into a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$upload(path, dest = NULL, progress = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the path to upload, absolute, or relative to
R's working directory.}

\item{\code{dest}}{Remote path save downloaded data to, relative to this
folder.  If \code{NULL} then the basename of the file is used.}

\item{\code{progress}}{Display httr's progress bar?}
}
\if{html}{\out{</div>}}
}
}
}
