% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\docType{package}
\name{spud}
\alias{spud}
\alias{spud-package}
\title{spud: R classes and methods for spatial usage data}
\description{
The spud package aims to give researchers and app developers an easy way to
visualize and analyze how and where people use a given service.
}
\section{Methods}{

Get started with spud using the \code{read.spud} method on the dummy data provided
with the package. Run \code{read.spud("dummy_data.csv")} and try out any of the
main exposed methods:

- plot_usage_actions_leaflet

- plot_usage_actions_mapview

- plot_user_path

Alternatively use the data to initialize a spud class.
}

\section{Classes}{

There are two classes included in spud, App and User, which represent an
application and its users, respectively. See the relevant documentation
pages for more details.
}

