% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{User}
\alias{User}
\title{User}
\description{
This class represents a user of a spatial app.
}
\section{Usage}{

\preformatted{
user = User$new("User 1", x)

user$first_action()
user$actions_map(flavor = "leaflet")
user$path_map()

print(user)
}
}

\section{Arguments}{

\describe{
  \item{id}{Character scalar, the id of the user.}
  \item{usage_data}{An sf object, representing the usage data of the user.}
}
}

\section{Details}{

\code{$new()} sets up the class with two fields: id and usage_data

\code{$print()} prints out some basic identifying information about the user.
}

\examples{
x = read.spud("dummy_data.csv")
user <- User$new("User 1", x)

}
