% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_client.R
\name{validate_cookies}
\alias{validate_cookies}
\title{Validate cookies in response}
\usage{
validate_cookies(response)
}
\arguments{
\item{response}{}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
To be able to use cookies in subsequent requests to sharepoint we
require the rtFa and FedAuth cookies to be set
}
\keyword{internal}
