% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{App}
\alias{App}
\title{App}
\description{
This class represents a spatial app and its associated usage data.
}
\section{Usage}{

\preformatted{
app = App$new("My app", x)

app$users()
app$user_count()
app$get_user(user_id)
app$actions_map(flavor = "leaflet")

print(app)
}
}

\section{Arguments}{

\describe{
  \item{name}{Character scalar, the name of the app.}
  \item{usage_data}{An sf object, representing the usage data for the app.}
}
}

\section{Details}{

\code{$new()} sets up the class with two fields: name and usage_data

\code{$print()} prints out some basic identifying information about the app.
}

\examples{
x = read.spu("dummy_data.csv")
app = App$new("My app", x)

}
