% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{plot_first_actions}
\alias{plot_first_actions}
\title{Draw map showing locations where users tried the app for the first time}
\usage{
plot_first_actions(data)
}
\arguments{
\item{data}{A spud object containing the data to be displayed}
}
\description{
Draw map showing locations where users tried the app for the first time
}
