% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{plot_usage_actions_leaflet}
\alias{plot_usage_actions_leaflet}
\title{Draw map showing spatial distribution of usage actions using leaflet}
\usage{
plot_usage_actions_leaflet(data)
}
\arguments{
\item{data}{A spud object containing the data to be displayed}
}
\description{
Draw map showing spatial distribution of usage actions using leaflet
}
