% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{plot_usage_actions_mapview}
\alias{plot_usage_actions_mapview}
\title{Draw map showing spatial distribution of usage actions using mapview}
\usage{
plot_usage_actions_mapview(data)
}
\arguments{
\item{data}{A spud object containing the data to be displayed}
}
\description{
Draw map showing spatial distribution of usage actions using mapview
}
