% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{plot_user_path}
\alias{plot_user_path}
\title{Draw map showing a path of a user over time}
\usage{
plot_user_path(data, user_id)
}
\arguments{
\item{data}{A spud object containing the data to be displayed}

\item{user_id}{The identifier of one user}
}
\description{
Draw map showing a path of a user over time
}
