% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spu.R
\name{read.spu}
\alias{read.spu}
\title{Read data from a csv file in the extdata directory.}
\usage{
read.spu(file, crs = 4326)
}
\arguments{
\item{file}{The name of the file you want to read in.}

\item{crs}{The coordinate reference system of the data.}
}
\value{
The resulting spu object.
}
\description{
Read data from a csv file in the extdata directory.
}
