% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spud.R
\name{sharepoint}
\alias{sharepoint}
\title{Create sharepoint connection for downloading data.}
\description{
Create sharepoint connection for downloading data.

Create sharepoint connection for downloading data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client}}{A low-level sharepoint client object, which can be used to interact
directly with the sharepoint API.  This object mostly handles
authentication, etc.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint$new()}}
\item \href{#method-download}{\code{sharepoint$download()}}
\item \href{#method-folder}{\code{sharepoint$folder()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create sharepoint object for downloading data from sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$new(sharepoint_url, auth = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_url}}{Root URL of sharepoint site to download from}

\item{\code{auth}}{Authentication data passed to the client}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{sharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download data from sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$download(
  sharepoint_path,
  dest = NULL,
  progress = FALSE,
  overwrite = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sharepoint_path}}{Path to the resource within sharepoint}

\item{\code{dest}}{Path to save downloaded data to}

\item{\code{progress}}{Display a progress bar during download?}

\item{\code{overwrite}}{Overwrite existing files?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Path to saved data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\if{latex}{\out{\hypertarget{method-folder}{}}}
\subsection{Method \code{folder()}}{
Create a \code{folder} object representing a sharepoint folder,
with which one can list, download and upload files.  See
\code{\link{sharepoint_folder}} for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint$folder(site, path, verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site}}{The name of the sharepoint site (most likely a short string)}

\item{\code{path}}{Relative path within that shared site.  It seems
that "Shared Documents" is a common path that most likely
represents a "Documents" collection when viewed in the
sharepoint web interface.}

\item{\code{verify}}{Logical, indicating if the site/path combination is
valid (slower but safer).}
}
\if{html}{\out{</div>}}
}
}
}
