% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_folder.R
\name{sharepoint_folder}
\alias{sharepoint_folder}
\title{Interact with sharepoint folders and their files.}
\description{
Interact with sharepoint folders and their files.

Interact with sharepoint folders and their files.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path of the folder (readonly)}

\item{\code{drive}}{A \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} object representing the
sharepoint drive.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint_folder$new()}}
\item \href{#method-files}{\code{sharepoint_folder$files()}}
\item \href{#method-folders}{\code{sharepoint_folder$folders()}}
\item \href{#method-list}{\code{sharepoint_folder$list()}}
\item \href{#method-delete}{\code{sharepoint_folder$delete()}}
\item \href{#method-parent}{\code{sharepoint_folder$parent()}}
\item \href{#method-folder}{\code{sharepoint_folder$folder()}}
\item \href{#method-create}{\code{sharepoint_folder$create()}}
\item \href{#method-download}{\code{sharepoint_folder$download()}}
\item \href{#method-upload}{\code{sharepoint_folder$upload()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create sharepoint_folder object to enable listing, creating
downloading and uploading files & folders
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$new(auth, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{auth}}{Auth info passed to \code{\link[Microsoft365R:client]{Microsoft365R::get_sharepoint_site()}}
contains site_name, site_url, site_id, tenant, app and scopes.}

\item{\code{path}}{Relative path within that shared site.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-files"></a>}}
\if{latex}{\out{\hypertarget{method-files}{}}}
\subsection{Method \code{files()}}{
List all files within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$files(path = "/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folders"></a>}}
\if{latex}{\out{\hypertarget{method-folders}{}}}
\subsection{Method \code{folders()}}{
List all folders within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folders(path = "/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
List all folders and files within the folder; this is a
convenience wrapper around the \code{files} and \code{folders} methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$list(path = "/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder, uses this folder if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a folder or file. Be extremely careful as you
could use this to delete an entire sharepoint. Deleted files are sent
to the recycle bin, so can be restored with relative ease, but
it will still be alarming. There is a mechanism to prevent
accidental deletion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$delete(path, check = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to delete. Use \code{NULL} to delete the current
folder.}

\item{\code{check}}{If TRUE then prompts user for confirmation before deleting}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parent"></a>}}
\if{latex}{\out{\hypertarget{method-parent}{}}}
\subsection{Method \code{parent()}}{
Create an object referring to the parent folder. If this
folder is the root, retrieving parent just returns the root again.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$parent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\if{latex}{\out{\hypertarget{method-folder}{}}}
\subsection{Method \code{folder()}}{
Create an object referring to a child folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folder(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the folder, relative to this folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create"></a>}}
\if{latex}{\out{\hypertarget{method-create}{}}}
\subsection{Method \code{create()}}{
Create a folder on sharepoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$create(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Folder relative to this folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download a file from a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$download(path, dest = NULL, overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path of the file to download, relative to this folder}

\item{\code{dest}}{Path to save downloaded data to. If \code{NULL} then a
temporary file with the same file extension as \code{path} is used.
If code{raw()} (or any other raw value) then the raw bytes will be
returned.}

\item{\code{overwrite}}{Overwrite the file if it exists?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload a file into a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$upload(path, dest = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the path to upload, absolute, or relative to
R's working directory.}

\item{\code{dest}}{Remote path save downloaded data to, relative to this
folder.  If \code{NULL} then uploaded to active folder.}
}
\if{html}{\out{</div>}}
}
}
}
