% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_folder.R
\name{sharepoint_folder}
\alias{sharepoint_folder}
\title{Interact with sharepoint folders and their files.}
\description{
Interact with sharepoint folders and their files.

Interact with sharepoint folders and their files.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sharepoint_folder$new()}}
\item \href{#method-files}{\code{sharepoint_folder$files()}}
\item \href{#method-folders}{\code{sharepoint_folder$folders()}}
\item \href{#method-list}{\code{sharepoint_folder$list()}}
\item \href{#method-parent}{\code{sharepoint_folder$parent()}}
\item \href{#method-folder}{\code{sharepoint_folder$folder()}}
\item \href{#method-download}{\code{sharepoint_folder$download()}}
\item \href{#method-upload}{\code{sharepoint_folder$upload()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$new(client, site, path, verify = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-files"></a>}}
\subsection{Method \code{files()}}{
List all files within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folders"></a>}}
\subsection{Method \code{folders()}}{
List all folders within the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\subsection{Method \code{list()}}{
List all folders and files within the folder; this is a
convenience wrapper around the \code{files} and \code{folders} methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parent"></a>}}
\subsection{Method \code{parent()}}{
Create an object referring to the parent folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$parent(verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verify}}{Verify that the folder exists (which it must really here)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-folder"></a>}}
\subsection{Method \code{folder()}}{
Create an object referring to a child folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$folder(path, verify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the folder, relative to this folder}

\item{\code{verify}}{Verify that the folder exists (which it must really here)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\subsection{Method \code{download()}}{
Download a file from a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$download(path, dest = NULL, progress = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the path to download, relative to this folder}

\item{\code{dest}}{Path to save downloaded data to. If \code{NULL} then a
temporary file with the same file extension as \code{path} is used.}

\item{\code{progress}}{Display httr's progress bar?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\subsection{Method \code{upload()}}{
Upload a file into a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sharepoint_folder$upload(path, dest = NULL, progress = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The name of the path to upload, absolute, or relative to
R's working directory.}

\item{\code{dest}}{Remote path save downloaded data to, relative to this
folder.  If \code{NULL} then the basename of the file is used.}

\item{\code{progress}}{Display httr's progress bar?}
}
\if{html}{\out{</div>}}
}
}
}
